feat: add new API method(s)
refactor: overhaul on code structure
- split structures into their own files
- enforce code style
- make define_variant macro more flexible

feat: add alloc feature guard

docs: cleanup documentation

chore: add rustfmt.toml
chore: modify rust.yml
chore: bump version to 0.3.0
